python/requirements.txt
semantic_kernel
pyyaml
pydantic
jinja2
pybars3
ruff
black